Version: 1.2.4 [2009-10-30] o ROBUSTIFICATION: Lowered the risk for saveObject() to leave an   imcomplete file due to say power failures etc.  This is done by   first writing to a temporary file, which is then renamed.  If the   temporary file already exists, an exception is thrown. o ROBUSTIFICATION: Now Arguments$getWritablePathname(path) validates   that there is enough file permissions so that a file can be created   in the 'path' directory. o CLEAN UP: On Windows Vista, createLink() produced a stderr message   "You do not have sufficient privilege to perform this operation", when   trying to use Windows 'mklink' command.  Those message are now silenced.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/r-dots/pkg/R.utils@302 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ R/saveObject.R
@@ -63,7 +63,7 @@ setMethodS3("saveObject", "default", function(object, file=NULL, path=NULL, comp
     # Write to a temporary file
     file <- pathnameT;
     on.exit({
-      if (file.exists(pathnameT)) {
+      if (!is.null(pathnameT) && file.exists(pathnameT)) {
         file.remove(pathnameT);
       }
     }, add=TRUE);
@@ -78,6 +78,7 @@ setMethodS3("saveObject", "default", function(object, file=NULL, path=NULL, comp
     if (!file.exists(pathname) || file.exists(pathnameT)) {
       throw("Failed to rename temporary file: ", pathnameT, " -> ", pathname);
     }
+    pathnameT <- NULL;
     file <- pathname;
   }
